Fix Windows warnings
src/cpu/quantities_cache/quantities_cache_windows.cpp: In function 'iware::cpu::cache_t iware::cpu::cache(unsigned int)':
src/cpu/quantities_cache/quantities_cache_windows.cpp:78:81: warning: 'type' may be used uninitialized in this function [-Wmaybe-uninitialized]
     return {info.Cache.Size, info.Cache.LineSize, info.Cache.Associativity, type};

src/system/OS_info/os_info_windows.cpp: In function 'std::__cxx11::string version_name()':
src/system/OS_info/os_info_windows.cpp:93:24: warning: overflow in conversion from 'tag_WBEM_TIMEOUT_TYPE' to 'LONG' {aka 'long int'} changes value from '(tag_WBEM_TIMEOUT_TYPE)4294967295' to '-1' [-Woverflow]
   query_iterator->Next(WBEM_INFINITE, 1, &value_raw, &iter_result);

+++ src/system/OS_info/os_info_windows.cpp
@@ -90,7 +90,7 @@ static std::string version_name() {
 		IWbemClassObject* value_raw;
 		unsigned long iter_result;
 
-		query_iterator->Next(WBEM_INFINITE, 1, &value_raw, &iter_result);
+		query_iterator->Next(static_cast<LONG>(WBEM_INFINITE), 1, &value_raw, &iter_result);
 		if(!iter_result)
 			break;
 		std::unique_ptr<IWbemClassObject, iware::detail::release_deleter> value(value_raw);
